+++ packages/lumx-react/src/components/image-block/ImageCaption.tsx
@@ -1,15 +1,22 @@
 import React, { CSSProperties, ReactNode } from 'react';
 
 import { FlexBox, HorizontalAlignment, Text, TextProps } from '@lumx/react';
-import { HasClassName, HasPolymorphicAs, HasTheme } from '@lumx/react/utils/type';
+import { HasPolymorphicAs, HasTheme } from '@lumx/react/utils/type';
+import classNames from 'classnames';
 
 type As = 'div' | 'figcaption';
+
+type ForwardedTextProps = Omit<TextProps, 'as' | 'typography' | 'color' | 'colorVariant'>;
 
 export type ImageCaptionMetadata = {
     /** Image title to display in the caption. */
     title?: string;
+    /** Props to pass to the title. */
+    titleProps?: ForwardedTextProps;
     /** Image description. Can be either a string, or sanitized html. */
     description?: string | { __html: string };
+    /** Props to pass to the title. */
+    descriptionProps?: ForwardedTextProps;
     /** Tag content. */
     tags?: ReactNode;
     /** Caption custom CSS style. */
@@ -17,9 +24,10 @@ export type ImageCaptionMetadata = {
 };
 
 export type ImageCaptionProps<AS extends As = 'figcaption'> = HasTheme &
-    HasClassName &
     HasPolymorphicAs<AS> &
     ImageCaptionMetadata & {
+        /** Base className for sub elements */
+        baseClassName?: string;
         /** Alignment. */
         align?: HorizontalAlignment;
         /** Truncate text on title & description (no line wrapping). */
@@ -28,7 +36,19 @@ export type ImageCaptionProps<AS extends As = 'figcaption'> = HasTheme &
 
 /** Internal component used to render image captions */
 export const ImageCaption = <AS extends As>(props: ImageCaptionProps<AS>) => {
-    const { className, theme, as = 'figcaption', title, description, tags, captionStyle, align, truncate } = props;
+    const {
+        baseClassName,
+        theme,
+        as = 'figcaption',
+        title,
+        titleProps,
+        description,
+        descriptionProps,
+        tags,
+        captionStyle,
+        align,
+        truncate,
+    } = props;
     if (!title && !description && !tags) return null;
 
     const titleColor = { color: theme === 'dark' ? 'light' : 'dark' } as const;
@@ -41,7 +61,7 @@ export const ImageCaption = <AS extends As>(props: ImageCaptionProps<AS>) => {
     return (
         <FlexBox
             as={as}
-            className={className}
+            className={classNames(baseClassName && `${baseClassName}__wrapper`)}
             style={captionStyle}
             orientation="vertical"
             vAlign={align}
@@ -49,17 +69,43 @@ export const ImageCaption = <AS extends As>(props: ImageCaptionProps<AS>) => {
             gap="regular"
         >
             {(title || description) && (
-                <Text as="p" truncate={truncate} {...baseColor}>
+                <Text
+                    as="p"
+                    className={classNames(baseClassName && `${baseClassName}__caption`)}
+                    truncate={truncate}
+                    {...baseColor}
+                >
                     {title && (
-                        <Text as="span" typography="subtitle1" {...titleColor}>
+                        <Text
+                            {...titleProps}
+                            as="span"
+                            className={classNames(titleProps?.className, baseClassName && `${baseClassName}__title`)}
+                            typography="subtitle1"
+                            {...titleColor}
+                        >
                             {title}
                         </Text>
                     )}{' '}
-                    {description && <Text as="span" typography="body1" {...descriptionContent} />}
+                    {description && (
+                        <Text
+                            {...descriptionProps}
+                            as="span"
+                            className={classNames(
+                                descriptionProps?.className,
+                                baseClassName && `${baseClassName}__description`,
+                            )}
+                            typography="body1"
+                            {...descriptionContent}
+                        />
+                    )}
                 </Text>
             )}
             {tags && (
-                <FlexBox orientation="horizontal" vAlign={align}>
+                <FlexBox
+                    className={classNames(baseClassName && `${baseClassName}__tags`)}
+                    orientation="horizontal"
+                    vAlign={align}
+                >
                     {tags}
                 </FlexBox>
             )}